+++ src/lungo_cli/core/plugin.py
@@ -22,14 +22,14 @@ from ..helpers.format import format_path, format_program
 from ..models.base import AppDirs, EApp
 from ..models.config import Config, Plugins, Privilege, Privileges, Rules
 from ..models.context import Context
-from ..models.plugin import BaseSettings, PluginConfig, PluginContext, PluginOutput
+from ..models.plugin import BaseSettings, PluginContext, PluginManifest, PluginOutput
 from ..models.users import Account, Extra, Users
 
 
 class BasePlugin[T: BaseSettings](ABC):
     """Base class for all plugins."""
 
-    config: ClassVar[PluginConfig]
+    manifest: ClassVar[PluginManifest]
 
     custom: ClassVar[bool] = False
     installable: ClassVar[bool] = False
@@ -78,18 +78,18 @@ class BasePlugin[T: BaseSettings](ABC):
         ip_addresses = self.context_manager.context.ip_addresses
 
         return PluginContext(
-            config=self.config,
-            backend_base_url=f"http://{ip_addresses[self.config.name]}:{self.config.backend_port}/",
+            manifest=self.manifest,
+            backend_base_url=f"http://{ip_addresses[self.manifest.name]}:{self.manifest.backend_port}/",
             dirs=AppDirs(
-                cache_dir=str(self.storage.cache_latest_dir / self.config.name),
-                generated_dir=str(self.storage.generated_dir / self.config.name),
-                managed_dir=str(self.storage.managed_dir / self.config.name),
-                plugin_dir=os.path.join(".", self.storage.plugins_rel / self.config.name),
+                cache_dir=str(self.storage.cache_latest_dir / self.manifest.name),
+                generated_dir=str(self.storage.generated_dir / self.manifest.name),
+                managed_dir=str(self.storage.managed_dir / self.manifest.name),
+                plugin_dir=os.path.join(".", self.storage.plugins_rel / self.manifest.name),
             ),
-            ip_address=ip_addresses[self.config.name],
-            oathkeeper_url_regex=f"{base_url}{PLUGIN_WEB_ENTRYPOINT}/<{self.config.web_path}>",
-            web_base_url=f"{base_url}{PLUGIN_WEB_ENTRYPOINT}/{self.config.web_path}/",
-            web_prefix=f"/{PLUGIN_WEB_ENTRYPOINT}/{self.config.web_path}",
+            ip_address=ip_addresses[self.manifest.name],
+            oathkeeper_url_regex=f"{base_url}{PLUGIN_WEB_ENTRYPOINT}/<{self.manifest.web_path}>",
+            web_base_url=f"{base_url}{PLUGIN_WEB_ENTRYPOINT}/{self.manifest.web_path}/",
+            web_prefix=f"/{PLUGIN_WEB_ENTRYPOINT}/{self.manifest.web_path}",
             custom=self.get_render_context(),
         )
 
@@ -98,10 +98,10 @@ class BasePlugin[T: BaseSettings](ABC):
     def output(self) -> PluginOutput:
         if self._output is None:
             if not self.rendered:
-                self.console.print_error(f"Plugin {format_program(self.config.name)} has not been rendered.")
+                self.console.print_error(f"Plugin {format_program(self.manifest.name)} has not been rendered.")
                 raise Exit(code=1)
 
-            plugin_dir = self.storage.installed_plugins_dir / self.config.name
+            plugin_dir = self.storage.installed_plugins_dir / self.manifest.name
 
             compose_content = self.file_utils.read_text(plugin_dir / "compose" / "compose.yaml", not_exist_ok=True)
             compose_services = extract_multiline_value_from_yaml(compose_content, "services")
@@ -115,7 +115,7 @@ class BasePlugin[T: BaseSettings](ABC):
             ).splitlines()
 
             self._output = PluginOutput(
-                config=self.config,
+                manifest=self.manifest,
                 compose_services=compose_services,
                 compose_secrets=compose_secrets,
                 nginx_site=nginx_site,
@@ -210,16 +210,16 @@ class PluginManager:
         """Add a plugin to the application."""
         if not plugin_cls.installable:
             self.console.print_warning(
-                f"Plugin {format_program(plugin_cls.config.name)} is not installable. Skipping it."
+                f"Plugin {format_program(plugin_cls.manifest.name)} is not installable. Skipping it."
             )
             return False
 
-        dst = self.storage.installed_plugins_dir / plugin_cls.config.name
+        dst = self.storage.installed_plugins_dir / plugin_cls.manifest.name
 
         if plugin_cls.custom:
-            self.file_utils.copy(self.storage.custom_plugins_dir / plugin_cls.config.name, dst)
+            self.file_utils.copy(self.storage.custom_plugins_dir / plugin_cls.manifest.name, dst)
         else:
-            self.file_utils.copy_package_resources(f"{PACKAGE_NAME}.plugins", plugin_cls.config.name, dst)
+            self.file_utils.copy_package_resources(f"{PACKAGE_NAME}.plugins", plugin_cls.manifest.name, dst)
 
         plugin_cls.installed = True
         return True
@@ -228,11 +228,11 @@ class PluginManager:
         """Remove a plugin from the application."""
         if not plugin_cls.installed:
             self.console.print_warning(
-                f"Plugin {format_program(plugin_cls.config.name)} is not installed. Skipping it."
+                f"Plugin {format_program(plugin_cls.manifest.name)} is not installed. Skipping it."
             )
             return False
 
-        self.file_utils.remove(self.storage.installed_plugins_dir / plugin_cls.config.name)
+        self.file_utils.remove(self.storage.installed_plugins_dir / plugin_cls.manifest.name)
 
         plugin_cls.installed = False
         return True
@@ -261,12 +261,12 @@ class PluginManager:
                 sys.path.insert(0, str(plugin_dir))
 
                 plugin_cls = import_module("plugin").Plugin
-                plugin_version = plugin_cls.config.version
+                plugin_version = plugin_cls.manifest.version
 
                 try:
-                    if get_app_version() not in SpecifierSet(plugin_cls.config.compatible_with):
+                    if get_app_version() not in SpecifierSet(plugin_cls.manifest.compatible_with):
                         self.console.print_warning(
-                            f"Plugin {format_program(plugin_cls.config.name)} "
+                            f"Plugin {format_program(plugin_cls.manifest.name)} "
                             f"{f'version {plugin_version}' if plugin_version else 'with an unknown version'} "
                             "is not compatible with the current version of the application. Skipping it."
                         )
@@ -274,7 +274,7 @@ class PluginManager:
                         plugin_cls.compatible = False
                 except InvalidSpecifier:
                     self.console.print_warning(
-                        f"Plugin {format_program(plugin_cls.config.name)} "
+                        f"Plugin {format_program(plugin_cls.manifest.name)} "
                         f"{f'version {plugin_version}' if plugin_version else 'with an unknown version'} "
                         "has an invalid version specifier. Skipping it."
                     )
@@ -296,14 +296,14 @@ class PluginManager:
         """Extend the models with fields for installed plugins."""
         for plugin_cls in self.compatible_plugin_classes:
             try:
-                extend_enum(EApp, plugin_cls.config.name.upper(), plugin_cls.config.name)
+                extend_enum(EApp, plugin_cls.manifest.name.upper(), plugin_cls.manifest.name)
 
                 settings_cls = plugin_cls.get_plugin_settings_cls()
                 field_info = FieldInfo(annotation=settings_cls, default=settings_cls())
-                Plugins.model_fields[plugin_cls.config.name] = field_info
+                Plugins.model_fields[plugin_cls.manifest.name] = field_info
             except Exception as e:
                 self.console.print_warning(
-                    f"Failed to load plugin {format_program(plugin_cls.config.name)} ({e}). Skipping it."
+                    f"Failed to load plugin {format_program(plugin_cls.manifest.name)} ({e}). Skipping it."
                 )
                 continue
 
@@ -320,13 +320,13 @@ class PluginManager:
     def initialize_plugins(self) -> None:
         """Initialize all plugins."""
         for plugin_cls in self.compatible_plugin_classes:
-            plugin_settings = getattr(self.context_manager.config.plugins, plugin_cls.config.name, None)
+            plugin_settings = getattr(self.context_manager.config.plugins, plugin_cls.manifest.name, None)
 
             if plugin_settings is None:
-                self.console.print_error(f"Settings for plugin {format_program(plugin_cls.config.name)} are missing.")
+                self.console.print_error(f"Settings for plugin {format_program(plugin_cls.manifest.name)} are missing.")
                 raise Exit(code=1)
 
             plugin = plugin_cls(plugin_settings, self.console, self.context_manager, self.file_utils, self.storage)
             plugin.update_data()
             self.plugins.append(plugin)
-            self.console.print_debug(f"Loaded plugin {format_program(plugin.config.name)}.")
+            self.console.print_debug(f"Loaded plugin {format_program(plugin.manifest.name)}.")