+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Breaking Changes
 
+- `Plugin.config` has been renamed to `Plugin.manifest`
 - `compatible_with` is now a required field for plugins
 - `version` is now a required field for plugins
 - Custom variables in the rendering context provided by plugins are now located under `plugin.custom`